refactor: Store 초기 상태, Reduce 함수 외부로 분리
